syncer: move UpstreamConn into dbconn package
