+++ syncer/syncer_test.go
@@ -29,6 +29,7 @@ import (
 	sqlmock "github.com/DATA-DOG/go-sqlmock"
 	"github.com/pingcap/failpoint"
 	"github.com/pingcap/parser/model"
+	"github.com/pingcap/tidb/infoschema"
 
 	"github.com/pingcap/dm/dm/config"
 	"github.com/pingcap/dm/dm/pb"
@@ -91,8 +92,11 @@ const (
 )
 
 type testSyncerSuite struct {
+	db              *sql.DB
 	cfg             *config.SubTaskConfig
 	eventsGenerator *event.Generator
+	syncer          *replication.BinlogSyncer
+	streamer        *replication.BinlogStreamer
 }
 
 type MockStreamer struct {
@@ -136,8 +140,8 @@ func (s *testSyncerSuite) SetUpSuite(c *C) {
 		Dir: loaderDir,
 	}
 	s.cfg = &config.SubTaskConfig{
-		From:         getDBConfigFromEnv(),
-		To:           getDBConfigFromEnv(),
+		From:         config.GetDBConfigFromEnv(),
+		To:           config.GetDBConfigFromEnv(),
 		ServerID:     101,
 		MetaSchema:   "test",
 		Name:         "syncer_ut",
@@ -150,7 +154,13 @@ func (s *testSyncerSuite) SetUpSuite(c *C) {
 
 	s.cfg.UseRelay = false
 
+	dbAddr := fmt.Sprintf("%s:%s@tcp(%s:%d)/?charset=utf8", s.cfg.From.User, s.cfg.From.Password, s.cfg.From.Host, s.cfg.From.Port)
+	s.db, err = sql.Open("mysql", dbAddr)
+	c.Assert(err, IsNil)
+
 	s.resetEventsGenerator(c)
+	_, err = s.db.Exec("SET GLOBAL binlog_format = 'ROW';")
+	c.Assert(err, IsNil)
 
 	c.Assert(log.InitLogger(&log.Config{}), IsNil)
 }
@@ -725,20 +735,14 @@ func (s *testSyncerSuite) TestColumnMapping(c *C) {
 
 func (s *testSyncerSuite) TestGeneratedColumn(c *C) {
 	// TODO Currently mock eventGenerator don't support generate json,varchar field event, so use real mysql binlog event here
-	dbAddr := fmt.Sprintf("%s:%s@tcp(%s:%d)/?charset=utf8mb4", s.cfg.From.User, s.cfg.From.Password, s.cfg.From.Host, s.cfg.From.Port)
-	db, err := sql.Open("mysql", dbAddr)
-	if err != nil {
-		c.Fatal(err)
-	}
-
-	_, err = db.Exec("SET GLOBAL binlog_format = 'ROW';")
+	_, err := s.db.Exec("SET GLOBAL binlog_format = 'ROW';")
 	c.Assert(err, IsNil)
 
-	pos, gset, err := utils.GetMasterStatus(context.Background(), db, "mysql")
+	pos, gset, err := utils.GetMasterStatus(context.Background(), s.db, "mysql")
 	c.Assert(err, IsNil)
 
 	//nolint:errcheck
-	defer db.Exec("drop database if exists gctest_1")
+	defer s.db.Exec("drop database if exists gctest_1")
 
 	s.cfg.BAList = &filter.Rules{
 		DoDBs: []string{"~^gctest_.*"},
@@ -869,16 +873,17 @@ func (s *testSyncerSuite) TestGeneratedColumn(c *C) {
 	}
 
 	for _, sql := range createSQLs {
-		_, err = db.Exec(sql)
+		_, err = s.db.Exec(sql)
 		c.Assert(err, IsNil)
 	}
 
 	syncer := NewSyncer(s.cfg, nil)
 	// use upstream dbConn as mock downstream
-	dbConn, err := db.Conn(context.Background())
+	dbConn, err := s.db.Conn(context.Background())
 	c.Assert(err, IsNil)
-	syncer.fromDB = &UpStreamConn{BaseDB: conn.NewBaseDB(db, func() {})}
-	syncer.ddlDB = syncer.fromDB.BaseDB
+	baseDB := conn.NewBaseDB(s.db, func() {})
+	syncer.fromDB = &dbconn.UpStreamConn{BaseDB: baseDB}
+	syncer.ddlDB = baseDB
 	syncer.ddlDBConn = &dbconn.DBConn{Cfg: s.cfg, BaseConn: conn.NewBaseConn(dbConn, &retry.FiniteRetryStrategy{})}
 	syncer.toDBConns = []*dbconn.DBConn{{Cfg: s.cfg, BaseConn: conn.NewBaseConn(dbConn, &retry.FiniteRetryStrategy{})}}
 	c.Assert(syncer.setSyncCfg(), IsNil)
@@ -892,7 +897,7 @@ func (s *testSyncerSuite) TestGeneratedColumn(c *C) {
 
 	for _, testCase := range testCases {
 		for _, sql := range testCase.sqls {
-			_, err = db.Exec(sql)
+			_, err = s.db.Exec(sql)
 			c.Assert(err, IsNil, Commentf("sql: %s", sql))
 		}
 		idx := 0
@@ -951,7 +956,7 @@ func (s *testSyncerSuite) TestGeneratedColumn(c *C) {
 	}
 
 	for _, sql := range dropSQLs {
-		_, err = db.Exec(sql)
+		_, err = s.db.Exec(sql)
 		c.Assert(err, IsNil)
 	}
 }
@@ -1054,7 +1059,7 @@ func (s *testSyncerSuite) TestRun(c *C) {
 
 	syncer := NewSyncer(s.cfg, nil)
 	syncer.cfg.CheckpointFlushInterval = 30
-	syncer.fromDB = &UpStreamConn{BaseDB: conn.NewBaseDB(db, func() {})}
+	syncer.fromDB = &dbconn.UpStreamConn{BaseDB: conn.NewBaseDB(db, func() {})}
 	syncer.toDBConns = []*dbconn.DBConn{
 		{Cfg: s.cfg, BaseConn: conn.NewBaseConn(dbConn, &retry.FiniteRetryStrategy{})},
 		{Cfg: s.cfg, BaseConn: conn.NewBaseConn(dbConn, &retry.FiniteRetryStrategy{})},
@@ -1295,7 +1300,7 @@ func (s *testSyncerSuite) TestExitSafeModeByConfig(c *C) {
 	}
 
 	syncer := NewSyncer(s.cfg, nil)
-	syncer.fromDB = &UpStreamConn{BaseDB: conn.NewBaseDB(db, func() {})}
+	syncer.fromDB = &dbconn.UpStreamConn{BaseDB: conn.NewBaseDB(db, func() {})}
 	syncer.toDBConns = []*dbconn.DBConn{
 		{Cfg: s.cfg, BaseConn: conn.NewBaseConn(dbConn, &retry.FiniteRetryStrategy{})},
 		{Cfg: s.cfg, BaseConn: conn.NewBaseConn(dbConn, &retry.FiniteRetryStrategy{})},
@@ -1838,4 +1843,177 @@ func (s *testSyncerSuite) TestDownstreamTableHasAutoRandom(c *C) {
 	ti.UpdateTS = ti2.UpdateTS
 
 	c.Assert(ti, DeepEquals, ti2)
+}
+
+func (s *testSyncerSuite) TestExecuteSQLSWithIgnore(c *C) {
+	db, mock, err := sqlmock.New()
+	c.Assert(err, IsNil)
+	dbConn, err := db.Conn(context.Background())
+	c.Assert(err, IsNil)
+	conn := &dbconn.DBConn{
+		BaseConn: &conn.BaseConn{
+			DBConn:        dbConn,
+			RetryStrategy: &retry.FiniteRetryStrategy{},
+		},
+		Cfg: &config.SubTaskConfig{
+			Name: "test",
+		},
+	}
+
+	sqls := []string{"alter table t1 add column a int", "alter table t1 add column b int"}
+
+	// will ignore the first error, and continue execute the second sql
+	mock.ExpectBegin()
+	mock.ExpectExec("alter table t1 add column a int").WillReturnError(newMysqlErr(uint16(infoschema.ErrColumnExists.Code()), "column a already exists"))
+	mock.ExpectExec("alter table t1 add column b int").WillReturnResult(sqlmock.NewResult(1, 1))
+	mock.ExpectCommit()
+
+	tctx := tcontext.Background().WithLogger(log.With(zap.String("test", "TestExecuteSQLSWithIgnore")))
+	n, err := conn.ExecuteSQLWithIgnore(tctx, ignoreDDLError, sqls)
+	c.Assert(err, IsNil)
+	c.Assert(n, Equals, 2)
+
+	// will return error when execute the first sql
+	mock.ExpectBegin()
+	mock.ExpectExec("alter table t1 add column a int").WillReturnError(newMysqlErr(uint16(infoschema.ErrColumnExists.Code()), "column a already exists"))
+	mock.ExpectRollback()
+
+	n, err = conn.ExecuteSQL(tctx, sqls)
+	c.Assert(err, ErrorMatches, ".*column a already exists.*")
+	c.Assert(n, Equals, 0)
+
+	c.Assert(mock.ExpectationsWereMet(), IsNil)
+}
+
+func (s *testSyncerSuite) TestTimezone(c *C) {
+	s.cfg.BAList = &filter.Rules{
+		DoDBs:     []string{"~^tztest_.*"},
+		IgnoreDBs: []string{"stest", "~^foo.*"},
+	}
+
+	createSQLs := []string{
+		"create database if not exists tztest_1",
+		"create table if not exists tztest_1.t_1(id int, a timestamp)",
+	}
+
+	testCases := []struct {
+		sqls     []string
+		timezone string
+	}{
+		{
+			[]string{
+				"insert into tztest_1.t_1(id, a) values (1, '1990-04-15 01:30:12')",
+				"insert into tztest_1.t_1(id, a) values (2, '1990-04-15 02:30:12')",
+				"insert into tztest_1.t_1(id, a) values (3, '1990-04-15 03:30:12')",
+			},
+			"Asia/Shanghai",
+		},
+		{
+			[]string{
+				"insert into tztest_1.t_1(id, a) values (4, '1990-04-15 01:30:12')",
+				"insert into tztest_1.t_1(id, a) values (5, '1990-04-15 02:30:12')",
+				"insert into tztest_1.t_1(id, a) values (6, '1990-04-15 03:30:12')",
+			},
+			"America/Phoenix",
+		},
+	}
+	queryTS := "select unix_timestamp(a) from `tztest_1`.`t_1` where id = ?"
+
+	dropSQLs := []string{
+		"drop table tztest_1.t_1",
+		"drop database tztest_1",
+	}
+
+	defer func() {
+		for _, sql := range dropSQLs {
+			_, err := s.db.Exec(sql)
+			c.Assert(err, IsNil)
+		}
+	}()
+
+	for _, sql := range createSQLs {
+		_, err := s.db.Exec(sql)
+		c.Assert(err, IsNil)
+	}
+
+	for _, testCase := range testCases {
+		syncer := NewSyncer(s.cfg, nil)
+		c.Assert(syncer.genRouter(), IsNil)
+		s.resetBinlogSyncer(c)
+
+		// we should not use `sql.DB.Exec` to do query which depends on session variables
+		// because `sql.DB.Exec` will choose a underlying DBConn for every query from the connection pool
+		// and different Conn using different session
+		// ref: `sql.DB.Conn`
+		// and `set @@global` is also not reasonable, because it can not affect sessions already exist
+		// if we must ensure multi queries use the same session, we should use a transaction
+		txn, err := s.db.Begin()
+		c.Assert(err, IsNil)
+		_, err = txn.Exec("set @@session.time_zone = ?", testCase.timezone)
+		c.Assert(err, IsNil)
+		_, err = txn.Exec("set @@session.sql_mode = ''")
+		c.Assert(err, IsNil)
+		for _, sql := range testCase.sqls {
+			_, err = txn.Exec(sql)
+			c.Assert(err, IsNil)
+		}
+		err = txn.Commit()
+		c.Assert(err, IsNil)
+
+		idx := 0
+		for {
+			if idx >= len(testCase.sqls) {
+				break
+			}
+			e, err := s.streamer.GetEvent(context.Background())
+			c.Assert(err, IsNil)
+			switch ev := e.Event.(type) {
+			case *replication.RowsEvent:
+				skip, err := syncer.skipDMLEvent(string(ev.Table.Schema), string(ev.Table.Table), e.Header.EventType)
+				c.Assert(err, IsNil)
+				if skip {
+					continue
+				}
+
+				rowid := ev.Rows[0][0].(int32)
+				var ts sql.NullInt64
+				err2 := s.db.QueryRow(queryTS, rowid).Scan(&ts)
+				c.Assert(err2, IsNil)
+				c.Assert(ts.Valid, IsTrue)
+
+				raw := ev.Rows[0][1].(string)
+				data, err := time.ParseInLocation("2006-01-02 15:04:05", raw, time.UTC)
+				c.Assert(err, IsNil)
+				c.Assert(data.Unix(), DeepEquals, ts.Int64)
+				idx++
+			default:
+				continue
+			}
+		}
+	}
+}
+
+func (s *testSyncerSuite) resetBinlogSyncer(c *C) {
+	cfg := replication.BinlogSyncerConfig{
+		ServerID:       s.cfg.ServerID,
+		Flavor:         "mysql",
+		Host:           s.cfg.From.Host,
+		Port:           uint16(s.cfg.From.Port),
+		User:           s.cfg.From.User,
+		Password:       s.cfg.From.Password,
+		UseDecimal:     false,
+		VerifyChecksum: true,
+	}
+	cfg.TimestampStringLocation = time.UTC
+
+	if s.syncer != nil {
+		s.syncer.Close()
+	}
+
+	pos, _, err := utils.GetMasterStatus(context.Background(), s.db, "mysql")
+	c.Assert(err, IsNil)
+
+	s.syncer = replication.NewBinlogSyncer(cfg)
+	s.streamer, err = s.syncer.StartSync(pos)
+	c.Assert(err, IsNil)
 }
